Restrict file operations in invites folder

+++ MailRuCloud.dpr
@@ -738,7 +738,7 @@ begin
 	Result := FS_FILE_NOTSUPPORTED;
 	If CheckFlag(FS_COPYFLAGS_RESUME, CopyFlags) then exit; {NEVER CALLED HERE}
 	RealPath := ExtractRealPath(RemoteName);
-	if RealPath.trashDir or RealPath.sharedDir then exit;
+	if RealPath.trashDir or RealPath.sharedDir or RealPath.invitesDir then exit;
 
 	MyProgressProc(PluginNum, RemoteName, LocalName, 0);
 
@@ -815,7 +815,7 @@ var
 begin
 
 	RealPath := ExtractRealPath(RemoteName);
-	if (RealPath.account = '') or RealPath.trashDir or RealPath.sharedDir then exit(FS_FILE_NOTSUPPORTED);
+	if (RealPath.account = '') or RealPath.trashDir or RealPath.sharedDir or RealPath.invitesDir then exit(FS_FILE_NOTSUPPORTED);
 	MyProgressProc(PluginNum, LocalName, pWideChar(RealPath.path), 0);
 
 	if CheckFlag(FS_COPYFLAGS_RESUME, CopyFlags) then exit(FS_FILE_NOTSUPPORTED); //NOT SUPPORTED
@@ -874,7 +874,7 @@ var
 	Invite: TCloudMailRuInviteInfo;
 Begin
 	RealPath := ExtractRealPath(WideString(RemoteName));
-	if (RealPath.account = '') or RealPath.trashDir then exit(false);
+	if (RealPath.account = '') or RealPath.trashDir or RealPath.invitesDir then exit(false);
 	Cloud := ConnectionManager.get(RealPath.account, getResult);
 	if RealPath.sharedDir then
 	begin
@@ -897,7 +897,7 @@ Begin
 	if SkipListRenMov then exit(false); //skip create directory if this flag set on
 
 	RealPath := ExtractRealPath(WideString(path));
-	if (RealPath.account = '') or RealPath.trashDir or RealPath.sharedDir then exit(false);
+	if (RealPath.account = '') or RealPath.trashDir or RealPath.sharedDir or RealPath.invitesDir then exit(false);
 	Result := ConnectionManager.get(RealPath.account, getResult).createDir(RealPath.path);
 end;
 
@@ -914,7 +914,7 @@ Begin
 		exit(false);
 	end;
 	RealPath := ExtractRealPath(WideString(RemoteName));
-	if RealPath.trashDir or RealPath.sharedDir then exit(false);
+	if RealPath.trashDir or RealPath.sharedDir or RealPath.invitesDir then exit(false);
 	Result := ConnectionManager.get(RealPath.account, getResult).removeDir(RealPath.path);
 end;
